feat(Plugin): Add plugin
Add a new plugin for exporting selected items to a Zip file
