+++ MiAZ/data/plugins/export2zip.plugin
@@ -1,0 +1,10 @@
+[Plugin]
+Module=export2zip
+Name=MiAZExport2ZIP
+Loader=Python3
+Description=Export selected items to a zip file
+Authors=Tomás Vírseda <tomasvirseda@gmail.com>
+Copyright=Copyright © 2023 Tomás Vírseda
+Website=http://github.com/t00m/MiAZ
+Help=http://github.com/t00m/MiAZ/README.adoc
+Version=0.1
